pyproject.toml
[project]
name = "foobara-py"
version = "0.2.0"
description = "Python implementation of Foobara command pattern with MCP integration"
readme = "README.md"
license = {text = "MPL-2.0"}
requires-python = ">=3.10"
authors = [
    {name = "Foobara Universe", email = "foobara@example.com"}
]
keywords = [
    "command-pattern",
    "mcp",
    "model-context-protocol",
    "foobara",
    "domain-driven-design",
    "pydantic",
    "fastapi",
    "ai-tools",
    "claude",
    "llm",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Mozilla Public License 2.0 (MPL 2.0)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Framework :: Pydantic :: 2",
]

dependencies = [
    "pydantic>=2.0",
    "argon2-cffi>=23.0",
    "jinja2>=3.0",
]

[project.optional-dependencies]
mcp = [
    "mcp>=1.0",
]
agent = [
    "instructor>=1.0",
    "anthropic>=0.30",
    "openai>=1.0",
]
http = [
    "fastapi>=0.100",
    "uvicorn>=0.20",
    "httpx>=0.25",
]
cli = [
    "typer>=0.9",
]
persistence = [
    "sqlalchemy>=2.0",
    "pyyaml>=6.0",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "mypy>=1.0",
    "ruff>=0.1",
]
all = [
    "foobara-py[mcp,agent,http,cli,persistence,dev]",
]

[project.urls]
Homepage = "https://github.com/foobara/foobara-py"
Documentation = "https://foobara-py.readthedocs.io"
Repository = "https://github.com/foobara/foobara-py"

[project.scripts]
foob = "foobara_py.cli.foob:main"
foobara-py = "foobara_py.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["foobara_py"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--cov=foobara_py --cov-report=html --cov-report=term-missing --cov-report=json"
markers = [
    "unit: Fast unit tests with no I/O",
    "integration: Tests requiring external resources",
    "slow: Tests taking >1 second",
    "persistence: Database/storage tests",
    "connectors: MCP/HTTP/CLI connector tests",
    "async: Async command tests",
]

[tool.coverage.run]
source = ["foobara_py"]
omit = [
    "foobara_py/_deprecated/*",
    "foobara_py/generators/templates/*",
    "tests/*",
    "*/__pycache__/*",
    "*/site-packages/*",
]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "def __str__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
show_missing = true
precision = 2
skip_covered = false
skip_empty = false

# Coverage thresholds
fail_under = 71.5

[tool.coverage.html]
directory = "htmlcov"

[tool.coverage.json]
output = "coverage.json"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.format]
quote-style = "double"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]
